# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Flora API
  version: 1.0.0
  description: Backend documentation for Flora project. Includes product pagination, filtering, order creation and demo authentication.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-flora-api.onrender.com
    description: Production server placeholder. Replace after deploy.
tags:
  - name: Products
    description: Product list, filtering, pagination and product details
  - name: Orders
    description: Order creation and order list
  - name: Auth
    description: Demo authentication endpoints
paths:
  /api/products:
    get:
      tags:
        - Products
      summary: Get paginated and filtered products
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 8
          description: Items per page
        - in: query
          name: search
          schema:
            type: string
            example: rose
          description: Search by product name or description
        - in: query
          name: category
          schema:
            type: string
            enum: ['', romantic, spring, premium, gift]
          description: Product category filter
        - in: query
          name: sort
          schema:
            type: string
            enum: [popular, price-asc, price-desc, name-asc]
            default: popular
          description: Sorting mode
      responses:
        '200':
          description: Product list received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductsResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by id
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
          example: spring-elegance
      responses:
        '200':
          description: Product received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/orders:
    get:
      tags:
        - Orders
      summary: Get demo order list
      responses:
        '200':
          description: Orders received successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  totalItems:
                    type: integer
                    example: 1
    post:
      tags:
        - Orders
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register demo user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Demo user registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login demo user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Demo user logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - category
        - image
        - description
      properties:
        id:
          type: string
          example: spring-elegance
        name:
          type: string
          example: Spring Elegance
        price:
          type: number
          example: 35
        category:
          type: string
          enum: [romantic, spring, premium, gift]
          example: spring
        popular:
          type: integer
          example: 100
        image:
          type: string
          example: /assets/spring-elegance-large.jpg
        description:
          type: string
          example: Fresh spring bouquet with soft pink and white flowers.
        bestseller:
          type: boolean
          example: true
    PaginatedProductsResponse:
      type: object
      required:
        - items
        - totalItems
        - totalPages
        - currentPage
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        totalItems:
          type: integer
          example: 15
        totalPages:
          type: integer
          example: 2
        currentPage:
          type: integer
          example: 1
    CreateOrderRequest:
      type: object
      required:
        - productId
        - name
        - phone
        - address
      properties:
        productId:
          type: string
          example: spring-elegance
        productName:
          type: string
          example: Spring Elegance
        quantity:
          type: integer
          minimum: 1
          example: 1
        name:
          type: string
          minLength: 2
          maxLength: 60
          example: Ann
        phone:
          type: string
          minLength: 6
          maxLength: 30
          example: '+1 (555) 123-4567'
        address:
          type: string
          minLength: 3
          maxLength: 160
          example: 456 Floral Ave, Sydney NSW 2000 AU
        message:
          type: string
          maxLength: 600
          example: Please deliver it in the morning.
    Order:
      allOf:
        - $ref: '#/components/schemas/CreateOrderRequest'
        - type: object
          required:
            - id
            - status
            - createdAt
          properties:
            id:
              type: string
              example: order_1720000000000
            status:
              type: string
              example: created
            createdAt:
              type: string
              format: date-time
              example: '2026-07-08T19:00:00.000Z'
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: Ann
        email:
          type: string
          format: email
          example: ann@example.com
        password:
          type: string
          format: password
          example: strong-password
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: ann@example.com
        password:
          type: string
          format: password
          example: strong-password
    AuthResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              example: demo-user
            name:
              type: string
              example: Ann
            email:
              type: string
              example: ann@example.com
        token:
          type: string
          example: demo-jwt-token
    Error:
      type: object
      properties:
        message:
          type: string
          example: Product not found
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: Validation error
        details:
          type: array
          items:
            type: string
          example:
            - '"name" is required'